real time notifications are ready
