+++ myapp/notif_scheduler.py
@@ -6,13 +6,31 @@ from myapp import stock_api
 import datetime
 from dateutil.relativedelta import relativedelta, FR
 from django.core.cache import cache
+from channels.layers import get_channel_layer
+from asgiref.sync import async_to_sync
 
 
 # When we already have a job running what to do? we can use max_instances keyword
-# TODO: Caching system
-# TODO: Event listener for exceptions
 # this class uses apsscheduler
 # further info can be found at https://apscheduler.readthedocs.io/en/stable/userguide.html
+def send_to_socket(user, company_symbol, description, time):
+    channel_layer = get_channel_layer()
+    data_to_socket = {
+        "symbol": company_symbol,
+        "description": description,
+        "time": time,
+    }
+
+    # Trigger message sent to group
+    async_to_sync(channel_layer.group_send)(
+        str(user.pk),  # Group Name, Should always be string
+        {
+            "type": "notify",  # custom function in consumer.py
+            "data": data_to_socket,
+        },  # message
+    )
+
+
 class NotificationsScheduler:
     def __init__(self):
         self.scheduler = BackgroundScheduler()
@@ -22,7 +40,7 @@ class NotificationsScheduler:
 
     def start(self):
         print('Started running the jobs')
-        self.scheduler.add_job(self.set_active_notifications, 'interval', hours=1, id=self.notifications_id)
+        self.scheduler.add_job(self.set_active_notifications, 'interval', seconds=30, id=self.notifications_id)
         self.scheduler.add_job(self.check_tracking_model, 'interval', hours=12, id=self.stock_tracking_id)
         self.scheduler.add_job(self.check_notifications_analyst, 'interval', days=1, id=self.analyst_rec_id)
 
@@ -62,7 +80,6 @@ class NotificationsScheduler:
             cache.set(symbol + 'analystRec', rec, (30 * 60))  # cache for 30 mins
             return rec
 
-
     @staticmethod
     def is_bigger(value, api_value):
         return api_value > value
@@ -82,7 +99,7 @@ class NotificationsScheduler:
         operators = {'bigger': self.is_bigger, 'lower': self.is_lower, 'equal': self.is_equal}
         for notification in notifications:
             time_now = now()
-            should_check = (((time_now - notification.last_checked).seconds // 60 % 60) >= 0)  # check if 10 mins passed
+            should_check = (((time_now - notification.last_checked).seconds // 60 % 60) >= 0)  # check if 0 mins passed
             if should_check:
                 user = notification.user
                 key = notification.operand
@@ -100,6 +117,7 @@ class NotificationsScheduler:
                     rn = ReadyNotification(user=user, description=description, company_symbol=symbol)
                     rn.save()
                     notification.delete()
+                    send_to_socket(user, rn.company_symbol, rn.description, str(rn.time))
                 else:
                     notification.last_checked = now()
                     notification.save()
@@ -223,9 +241,3 @@ class NotificationsScheduler:
                 else:
                     notification.last_checked = now()
                     notification.save()
-    # def job_listener(self, event):
-    #     if event.exception:
-    #         print('The job crashed :(')
-    #         raise event.exception
-    #     else:
-    #         print('The job worked :)')
