+++ myapp/templates/popup_modal/notif_toast.html
@@ -1,12 +1,12 @@
 <div id="notif-toast" class="toast" role="alert" aria-live="assertive" aria-atomic="true" 
-style="position: absolute; top: 3%; right: 2%; z-index: 1; width: 300px;">
+style="position: absolute; top: 3%; right: 2%; z-index: 1; width: 300px; pointer-events: none;">
     <div class="toast-header">
         <!-- <img src="..." class="rounded mr-2" alt="..."> -->
-        <strong class="mr-auto">Notification status</strong>
+        <strong id="toast-title" class="mr-auto">Notification status</strong>
         <!-- <small class="text-muted">just now</small> -->
-        <button type="button" class="ml-2 mb-1 close" data-dismiss="toast" aria-label="Close">
+        <!-- <button type="button" class="ml-2 mb-1 close" data-dismiss="toast" aria-label="Close">
             <span aria-hidden="true">&times;</span>
-        </button>
+        </button> -->
     </div>
     <div id="toast-message" class="toast-body">
