+++ myapp/tests/test_notifications.py
@@ -1,8 +1,7 @@
 from django.test import TestCase, Client
-from django.contrib.auth.models import User
 from django.test import RequestFactory
-from django.urls import reverse
 from myapp.views import *
+from myapp.notifications_ajax import *
 import json
 from myapp.models import Company
