+++ myapp/views.py
@@ -1,35 +1,15 @@
-from django.shortcuts import render, redirect
-from myapp import stock_api
-from django.core.paginator import Paginator
-from django.http import JsonResponse
-from django.conf import settings
 from django.http import QueryDict
 from django.urls import reverse
-from myapp.models import Stock, UserProfile, UserStock, StockOperation, FavoriteStock
-from django.http import JsonResponse
-from django.contrib.auth.models import User
 from django.conf import settings
 from django.contrib.auth import logout
 from django.core import serializers
-# import wikipedia as wiki
 from django.core.files.storage import FileSystemStorage
 import pathlib
-from dateutil.relativedelta import relativedelta, MO
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
 from django.shortcuts import render, redirect, HttpResponse
 from myapp import stock_api
 from myapp.models import *
-
-HTTP_200_OK = 200
-HTTP_201_CREATED = 201
-HTTP_202_ACCEPTED = 202
-HTTP_204_NO_CONTENT = 204
-HTTP_304_NOT_MODIFIED = 304
-HTTP_400_BAD_REQUEST = 400
-HTTP_401_UNAUTHORIZED = 401
-HTTP_403_FORBIDDEN = 403
-HTTP_500_INTERNAL_SERVER_ERROR = 500
 
 
 # View for the home page - a list of 20 of the most active stocks
@@ -70,15 +50,16 @@ def compare(request):
 
 
 def exchange(request):
-	if not request.user.is_authenticated:
-		return redirect(reverse('login'))
-	user = request.user
-	user_stocks = UserStock.objects.get(user_id=user.id)
-	stocks = Stock.objects.filter(top_rank__isnull=False).order_by('top_rank')
-	return render(request, 'exchange.html', 
-                    {'user_stocks': serializers.serialize('json', user_stocks.stock_buyied.all()), 
-                    'stocks': serializers.serialize('json', stocks),
-                    'user_budget': user_stocks.budget})
+    if not request.user.is_authenticated:
+        return redirect(reverse('login'))
+    user = request.user
+    user_stocks = UserStock.objects.get(user_id=user.id)
+    stocks = Stock.objects.filter(top_rank__isnull=False).order_by('top_rank')
+    return render(request, 'exchange.html',
+                  {'user_stocks': serializers.serialize('json', user_stocks.stock_buyied.all()),
+                   'stocks': serializers.serialize('json', stocks),
+                   'user_budget': user_stocks.budget})
+
 
 # View for the single stock page
 # symbol is the requested stock's symbol ('AAPL' for Apple)
@@ -251,57 +232,6 @@ def logout_view(request):
     return redirect('index')
 
 
-# API for a stock's price over time
-# symbol is the requested stock's symbol ('AAPL' for Apple)
-# The response is JSON data of an array composed of "snapshot" objects (date + stock info + ...), usually one per day
-def single_stock_historic(request, symbol, time_range='1m'):
-    data = stock_api.get_stock_historic_prices(symbol, time_range=time_range)
-    return JsonResponse({'data': data})
-
-
-def get_company_desc(request, company_symbol):
-    try:
-        comp = Company.objects.get(company_symbol=company_symbol)
-        return JsonResponse({'summary': comp.company_desc})
-    except Exception as e:
-        print(e)
-        return JsonResponse({'summary': 'Couldn\'t find information'})
-
-
-@login_required
-def multi_stocks_historic(request, stocks, time_range='1m'):
-    result = []
-    stocks_list = stocks.split('-')
-
-    for stock in stocks_list:
-        result.append({'name': stock, 'data': stock_api.get_stock_historic_prices(stock, time_range=time_range)})
-
-    return JsonResponse({'data': result}, content_type="application/json")
-
-
-def stocks_names_and_symbols(request):
-    data = Stock.objects.filter(top_rank__isnull=False).order_by('name')
-    result = []
-    for datum in data:
-        result.append('{} - {}'.format(datum.symbol, datum.name))
-    return JsonResponse({'data': result}, content_type="application/json")
-
-
-def add_notification(request):
-    print('received')
-    if request.method == "POST":
-        print('posting', request.POST)
-        if request.user.is_authenticated:
-            notification = Notification(user=request.user, operator=request.POST.get('operator').strip()
-                                        , operand=request.POST.get('operand').strip()
-                                        , value=request.POST.get('value').strip()
-                                        , company_symbol=request.POST.get('company_symbol').strip())
-
-            notification.save()
-
-    return HttpResponse(status=HTTP_204_NO_CONTENT)
-
-
 @login_required
 def my_notifications(request):
     waiting_notifs = Notification.objects.filter(user=request.user)
@@ -340,51 +270,6 @@ def delete_waiting_notification(request, pk='-1'):
     return redirect('my_notifications')
 
 
-# TODO: make datetime zone aware
-def get_time(include_this_week):
-    # Mon 0 - Sun 6
-    today = datetime.datetime.now().weekday()
-    if today == 0:  # if today is Monday return the date
-        return now().date()
-
-    if include_this_week == '1':
-        if today == 5:  # if today is Saturday
-            return (now() + relativedelta(weekday=MO(+1))).date()  # return date of next Monday
-        elif today == 6:  # if today is Sunday
-            return (now() + relativedelta(weekday=MO(+1))).date()  # return date of next Monday
-        else:
-            return (now() + relativedelta(weekday=MO(-1))).date()  # return date of prev Monday
-    else:  # go to next week
-        return (now() + relativedelta(weekday=MO(+1))).date()  # return date of next Monday
-
-
-# TODO: show feedback to user
-@login_required
-def add_tracking(request):
-    if request.method == "POST":
-        if request.user.is_authenticated:
-            post = request.POST
-            track = TrackStock(user=request.user, operand=int(post.get('operand')), state=int(post.get('state'))
-                               , weeks=int(post.get('weeks')), company_symbol=post.get('company_symbol'),
-                               creation_time=get_time(post.get('include_this_week')))
-            track.save()
-
-    return HttpResponse(status=HTTP_204_NO_CONTENT)
-
-
-@login_required
-def add_notification_analyst(request):
-    if request.method == "POST":
-        if request.user.is_authenticated:
-            post = request.POST
-            notif = NotificationAnalystRec(user=request.user, operator=post.get('operator').strip()
-                                           , value=post.get('value').strip()
-                                           , company_symbol=post.get('company_symbol').strip())
-            notif.save()
-
-    return HttpResponse(status=HTTP_204_NO_CONTENT)
-
-
 @login_required
 def delete_tracking_notification(request, pk='-1'):
     if request.user.is_authenticated:
@@ -405,3 +290,16 @@ def delete_analyst_notification(request, pk='-1'):
                 n.delete()
 
     return redirect('my_notifications')
+
+# def get_top_five_notifs(request):
+#     number_of_notifs = 0
+#     notifications = []
+#     if request.user.is_authenticated:
+#         notifications = ReadyNotification.objects.filter(user=request.user).values('company_symbol',
+#                                                                                    'description',
+#                                                                                    'time').order_by('-id')
+#         number_of_notifs = len(notifications)
+#         notifications = notifications[:5]  # only recent 5
+#     return JsonResponse(
+#         {'number_of_notifs': number_of_notifs, 'notifications':     notifications},
+#         content_type="application/json")
